+++ py-polars/polars/expr/expr.py
@@ -21,8 +21,8 @@ from typing import (
     cast,
 )
 
+from polars import _reexport as pli
 from polars import functions as F
-from polars import internals as pli
 from polars.datatypes import (
     FLOAT_DTYPES,
     INTEGER_DTYPES,
