+++ py-polars/polars/io/parquet/anonymous_scan.py
@@ -4,7 +4,7 @@ from functools import partial
 from typing import TYPE_CHECKING, Any
 
 import polars.io.parquet
-from polars import internals as pli
+from polars import _reexport as pli
 from polars.dependencies import pickle
 from polars.io._utils import _prepare_file_arg
